Add _form partial for create and edit templates

+++ src/Controller/PinsController.php
@@ -66,10 +66,13 @@ class PinsController extends AbstractController
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"DELETE"})
      * @return Response
      */
-    public function delete(Pin $pin, EntityManagerInterface $em): Response
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
     {
-        $em->remove($pin);
-        $em->flush();
+        if ($this->isCsrfTokenValid('yeah' . $pin->getId(), $request->request->get('comedown'))) {
+            # code...
+            $em->remove($pin);
+            $em->flush();
+        }
         return $this->redirectToRoute('app_home');
     }
